refactor: header.js에 menu.js추가
