+++ src/Component/Header/Header.css
@@ -1,12 +1,15 @@
 .HeaderContainer {
-  width: 260px;
+  width: 320px;
   height: 100px;
   display: flex;
   justify-content: space-between;
+  background-image: url("../../asset/top.tiff");
+  background-size: cover;
+  background-position: center;
   position: fixed;
   z-index: 10;
   align-items: center;
-  margin-left: 100px;
+  margin-left: 50px;
   cursor: pointer;
 }
 
@@ -43,4 +46,13 @@
 }
 
 @media (max-width: 1024px) and (min-width: 769px) {
+}
+
+
+.HeaderTitleAndLogo {
+  width: 260px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  
 }
