+++ src/Component/HeaderMenu/HeaderMenu.js
@@ -1,0 +1,14 @@
+import "./HeaderMenu.css";
+import React from "react";
+
+const HeaderMenu = () => {
+  return (
+    <div className="HeaderMenu">
+      <div className="HaederMenuBar"></div>
+      <div className="HaederMenuBar"></div>
+      <div className="HaederMenuBar"></div>
+    </div>
+  );
+};
+
+export default HeaderMenu;
